+++ csvkitcat/moreutils/csvsed.py
@@ -4,9 +4,9 @@
 
 from csvkitcat.alltext import AllTextUtility
 from csvkit.grep import FilteringCSVReader
+from csvkitcat import rxlib as re
 
 import agate
-import regex as re
 import warnings
 
 from sys import stderr
@@ -51,60 +51,60 @@ class CSVSed(AllTextUtility):
                                     help='The CSV file to operate on. If omitted, will accept input as piped data via STDIN.')
 
     #     ###########boilerplate
-    #     self.argparser.add_argument('-D', '--out-delimiter', dest='out_delimiter',
-    #                                 help='Delimiting character of the output CSV file.')
-    #     self.argparser.add_argument('-T', '--out-tabs', dest='out_tabs', action='store_true',
-    #                                 help='Specify that the output CSV file is delimited with tabs. Overrides "-D".')
-    #     self.argparser.add_argument('-Q', '--out-quotechar', dest='out_quotechar',
-    #                                 help='Character used to quote strings in the output CSV file.')
-    #     self.argparser.add_argument('-U', '--out-quoting', dest='out_quoting', type=int, choices=[0, 1, 2, 3],
-    #                                 help='Quoting style used in the output CSV file. 0 = Quote Minimal, 1 = Quote All, 2 = Quote Non-numeric, 3 = Quote None.')
-    #     self.argparser.add_argument('-B', '--out-no-doublequote', dest='out_doublequote', action='store_false',
-    #                                 help='Whether or not double quotes are doubled in the output CSV file.')
-    #     self.argparser.add_argument('-P', '--out-escapechar', dest='out_escapechar',
-    #                                 help='Character used to escape the delimiter in the output CSV file if --quoting 3 ("Quote None") is specified and to escape the QUOTECHAR if --no-doublequote is specified.')
-    #     self.argparser.add_argument('-M', '--out-lineterminator', dest='out_lineterminator',
-    #                                 help='Character used to terminate lines in the output CSV file.')
+        self.argparser.add_argument('-D', '--out-delimiter', dest='out_delimiter',
+                                    help='Delimiting character of the output CSV file.')
+        self.argparser.add_argument('-T', '--out-tabs', dest='out_tabs', action='store_true',
+                                    help='Specify that the output CSV file is delimited with tabs. Overrides "-D".')
+        self.argparser.add_argument('-Q', '--out-quotechar', dest='out_quotechar',
+                                    help='Character used to quote strings in the output CSV file.')
+        self.argparser.add_argument('-U', '--out-quoting', dest='out_quoting', type=int, choices=[0, 1, 2, 3],
+                                    help='Quoting style used in the output CSV file. 0 = Quote Minimal, 1 = Quote All, 2 = Quote Non-numeric, 3 = Quote None.')
+        self.argparser.add_argument('-B', '--out-no-doublequote', dest='out_doublequote', action='store_false',
+                                    help='Whether or not double quotes are doubled in the output CSV file.')
+        self.argparser.add_argument('-P', '--out-escapechar', dest='out_escapechar',
+                                    help='Character used to escape the delimiter in the output CSV file if --quoting 3 ("Quote None") is specified and to escape the QUOTECHAR if --no-doublequote is specified.')
+        self.argparser.add_argument('-M', '--out-lineterminator', dest='out_lineterminator',
+                                    help='Character used to terminate lines in the output CSV file.')
 
 
-    # def _extract_csv_writer_kwargs(self):
-    #     kwargs = {}
+    def _extract_csv_writer_kwargs(self):
+        kwargs = {}
 
-    #     # if self.args.line_numbers:
-    #     #     kwargs['line_numbers'] = True
+        # if self.args.line_numbers:
+        #     kwargs['line_numbers'] = True
 
-    #     if self.args.out_tabs:
-    #         kwargs['delimiter'] = '\t'
-    #     elif self.args.out_delimiter:
-    #         kwargs['delimiter'] = self.args.out_delimiter
-
-
-    #     for arg in ('quotechar', 'quoting', 'doublequote', 'escapechar', 'lineterminator'):
-    #         value = getattr(self.args, 'out_%s' % arg)
-    #         if value is not None:
-    #             kwargs[arg] = value
-
-    #     return kwargs
+        if self.args.out_tabs:
+            kwargs['delimiter'] = '\t'
+        elif self.args.out_delimiter:
+            kwargs['delimiter'] = self.args.out_delimiter
 
 
-    # def run(self):
-    #     """
-    #     A wrapper around the main loop of the utility which handles opening and
-    #     closing files.
+        for arg in ('quotechar', 'quoting', 'doublequote', 'escapechar', 'lineterminator'):
+            value = getattr(self.args, 'out_%s' % arg)
+            if value is not None:
+                kwargs[arg] = value
 
-    #     TK: This is copy-pasted form CSVKitUtil because we have to override 'f'; maybe there's
-    #         a way to refactor this...
-    #     """
-    #     self.input_file = self._open_input_file(self.args.input_path)
+        return kwargs
 
-    #     try:
-    #         with warnings.catch_warnings():
-    #             if getattr(self.args, 'no_header_row', None):
-    #                 warnings.filterwarnings(action='ignore', message='Column names not specified', module='agate')
 
-    #             self.main()
-    #     finally:
-    #         self.input_file.close()
+    def run(self):
+        """
+        A wrapper around the main loop of the utility which handles opening and
+        closing files.
+
+        TK: This is copy-pasted form CSVKitUtil because we have to override 'f'; maybe there's
+            a way to refactor this...
+        """
+        self.input_file = self._open_input_file(self.args.input_path)
+
+        try:
+            with warnings.catch_warnings():
+                if getattr(self.args, 'no_header_row', None):
+                    warnings.filterwarnings(action='ignore', message='Column names not specified', module='agate')
+
+                self.main()
+        finally:
+            self.input_file.close()
 
     def main(self):
         if self.additional_input_expected():
@@ -132,7 +132,6 @@ class CSVSed(AllTextUtility):
         patterns = dict((column_id, pattern) for column_id in column_ids)
 
         if self.args.exclude_unmatched:
-            stderr.write(f"FILTERING\n")
             xreader = FilteringCSVReader(rows, header=False, patterns=patterns, inverse=False, any_match=True)
         else:
             xreader = rows
@@ -141,17 +140,7 @@ class CSVSed(AllTextUtility):
         output = agate.csv.writer(self.output_file, **writer_kwargs)
         output.writerow(column_names)
 
-
-        #### BENCHMARKING ISSUES
-        ####
-        # TK TODO THIS IS UGLY!!
-        z = 0 # TK: becnchmark code to delete later
-        # xreader = list(xreader)
-        # stderr.write(f"How many rows: {len(xreader)}\n")
-
         for z, row in enumerate(xreader):
-
-
             d = []
             for _x, val in enumerate(row):
                 newval = val
@@ -162,13 +151,12 @@ class CSVSed(AllTextUtility):
                         else:
                             mx = pattern.search(val)
                             if mx:
-                                newval = pattern.sub(repl, mx.captures(0)[0])
+                                newval = pattern.sub(repl, mx.group(0))
                     else:
                         newval = val.replace(pattern, repl, max_match_count) if self.args.literal_match else pattern.sub(repl, val, max_match_count)
                 d.append(newval)
             output.writerow(d)
 
-        # stderr.write(f"Z count: {z}\n")
 
 def launch_new_instance():
     utility = CSVSed()